Read page content from DB
