+++ src/server/action.ts
@@ -74,25 +74,15 @@ export async function getBookForHome() {
 export async function getBookById(bookId: string) {
   const data = await fetch(`http://localhost:8080/api/books/getBookById/${bookId}`)
   const response = await data.json()
-  // const selectedBook: booksFromServer = {
-  //   bookId: response.bookId,
-  //   title: response.title,
-  //   description: response.description,
-  //   author: response.author,
-  //   publisher: response.publisher,
-  //   bookCover: response.bookCover,
-  //   totalPages: response.totalPages,
-  //   categories: response.categories,
-  //   cloudUrl: response.cloudUrl,
-  //   premium: response.premium
-  // }
   let selectedBook: Array<booksFromServer> = []
   selectedBook.push(response)
   return selectedBook[0]
 }
 
-export async function getBookContentByPage() {
-  
+export async function getBookContentByPage(id: string, pageNum: number) {
+  const data = await fetch(`http://localhost:8080/api/books/getBookPage/${id}/pageNumber${pageNum}`)
+  const response = await data.text()
+  return response
 }
 
 export async function registerNewBook() {
